Renamed scripts to autogpt. Absolute Imports. isort. Fixed flake8 F401, E402
