+++ autogpt/agent_manager.py
@@ -1,4 +1,4 @@
-from llm_utils import create_chat_completion
+from autogpt.llm_utils import create_chat_completion
 
 next_key = 0
 agents = {}  # key, (task, full_message_history, model)
