+++ autogpt/ai_config.py
@@ -1,5 +1,6 @@
 import yaml
-import data
+
+from autogpt import data
 
 
 class AIConfig:
