+++ autogpt/call_ai_function.py
@@ -1,7 +1,9 @@
-from config import Config
+from autogpt.config import Config
+from autogpt.llm_utils import create_chat_completion
+
 cfg = Config()
 
-from llm_utils import create_chat_completion
+
 
 # This is a magic function that can do anything with no-code. See
 # https://github.com/Torantulino/AI-Functions for more info.
