+++ autogpt/chat.py
@@ -1,12 +1,13 @@
 import time
+
 import openai
-from dotenv import load_dotenv
-from config import Config
-import token_counter
+
+from autogpt import token_counter
+from autogpt.config import Config
+from autogpt.llm_utils import create_chat_completion
 
 cfg = Config()
 
-from llm_utils import create_chat_completion
 
 
 def create_chat_message(role, content):
@@ -56,7 +57,7 @@ def chat_with_ai(
             current_context = [
                 create_chat_message(
                     "system", prompt), create_chat_message(
-                    "system", f"Permanent memory: {permanent_memory}")]                
+                    "system", f"Permanent memory: {permanent_memory}")]
 
             # Add messages from the full message history until we reach the token limit
             next_message_to_add_index = len(full_message_history) - 1
@@ -80,7 +81,7 @@ def chat_with_ai(
 
                 # Count the currently used tokens
                 current_tokens_used += tokens_to_add
-                
+
                 # Move to the next most recent message in the full message history
                 next_message_to_add_index -= 1
