+++ autogpt/commands.py
@@ -1,17 +1,18 @@
-import browse
-import json
-import memory as mem
 import datetime
+import json
+
 import agent_manager as agents
+import browse
+import memory as mem
 import speak
-from config import Config
-import ai_functions as ai
-from file_operations import read_file, write_to_file, append_to_file, delete_file
-from execute_code import execute_python_file
-from json_parser import fix_and_parse_json
 from googlesearch import search
-from googleapiclient.discovery import build
-from googleapiclient.errors import HttpError
+
+import autogpt.ai_functions as ai
+from autogpt.config import Config
+from autogpt.execute_code import execute_python_file
+from autogpt.file_operations import (append_to_file, delete_file, read_file,
+                                     write_to_file)
+from autogpt.json_parser import fix_and_parse_json
 
 cfg = Config()
 
@@ -19,15 +20,15 @@ cfg = Config()
 def get_command(response):
     try:
         response_json = fix_and_parse_json(response)
-        
+
         if "command" not in response_json:
             return "Error:" , "Missing 'command' object in JSON"
-        
+
         command = response_json["command"]
 
         if "name" not in command:
             return "Error:", "Missing 'name' field in 'command' object"
-        
+
         command_name = command["name"]
 
         # Use an empty dictionary if 'args' field is not present in 'command' object
@@ -47,7 +48,7 @@ def get_command(response):
 def execute_command(command_name, arguments):
     try:
         if command_name == "google":
-            
+
             # Check if the Google API key is set and use the official search method
             # If the API key is not set or has only whitespaces, use the unofficial search method
             if cfg.google_api_key and (cfg.google_api_key.strip() if cfg.google_api_key else None):
@@ -118,9 +119,10 @@ def google_search(query, num_results=8):
     return json.dumps(search_results, ensure_ascii=False, indent=4)
 
 def google_official_search(query, num_results=8):
+    import json
+
     from googleapiclient.discovery import build
     from googleapiclient.errors import HttpError
-    import json
 
     try:
         # Get the Google API key and Custom Search Engine ID from the config file
@@ -129,20 +131,20 @@ def google_official_search(query, num_results=8):
 
         # Initialize the Custom Search API service
         service = build("customsearch", "v1", developerKey=api_key)
-        
+
         # Send the search query and retrieve the results
         result = service.cse().list(q=query, cx=custom_search_engine_id, num=num_results).execute()
 
         # Extract the search result items from the response
         search_results = result.get("items", [])
-        
+
         # Create a list of only the URLs from the search results
         search_results_links = [item["link"] for item in search_results]
 
     except HttpError as e:
         # Handle errors in the API call
         error_details = json.loads(e.content.decode())
-        
+
         # Check if the error is related to an invalid or missing API key
         if error_details.get("error", {}).get("code") == 403 and "invalid API key" in error_details.get("error", {}).get("message", ""):
             return "Error: The provided Google API key is invalid or missing."
