+++ autogpt/config.py
@@ -1,6 +1,8 @@
 import os
+
 import openai
 from dotenv import load_dotenv
+
 # Load environment variables from .env file
 load_dotenv()
