+++ autogpt/json_parser.py
@@ -1,6 +1,8 @@
 import json
-from call_ai_function import call_ai_function
-from config import Config
+
+from autogpt.call_ai_function import call_ai_function
+from autogpt.config import Config
+
 cfg = Config()
 
 def fix_and_parse_json(json_str: str, try_to_fix_with_gpt: bool = True):
@@ -47,7 +49,7 @@ def fix_and_parse_json(json_str: str, try_to_fix_with_gpt: bool = True):
               return json_str
           else:
             raise e
-        
+
 def fix_json(json_str: str, schema: str, debug=False) -> str:
     # Try to fix the JSON using gpt:
     function_string = "def fix_json(json_str: str, schema:str=None) -> str:"
