+++ autogpt/llm_utils.py
@@ -1,5 +1,7 @@
 import openai
-from config import Config
+
+from autogpt.config import Config
+
 cfg = Config()
 
 openai.api_key = cfg.openai_api_key
