+++ autogpt/main.py
@@ -1,21 +1,19 @@
+import argparse
 import json
 import random
-import commands as cmd
-import memory as mem
-import data
-import chat
-from colorama import Fore, Style
-from spinner import Spinner
 import time
-import speak
-from enum import Enum, auto
-import sys
-from config import Config
-from json_parser import fix_and_parse_json
-from ai_config import AIConfig
 import traceback
+
 import yaml
-import argparse
+from colorama import Fore, Style
+from config import Config
+
+import autogpt.commands as cmd
+import autogpt.memory as mem
+from autogpt import chat, data, speak
+from autogpt.ai_config import AIConfig
+from autogpt.json_parser import fix_and_parse_json
+from autogpt.spinner import Spinner
 
 
 def print_to_console(
@@ -115,7 +113,7 @@ def load_variables(config_file="config.yaml"):
         if ai_name == "":
             ai_name = "Entrepreneur-GPT"
 
-    if not ai_role:        
+    if not ai_role:
         ai_role = input(f"{ai_name} is: ")
         if ai_role == "":
             ai_role = "an AI designed to autonomously develop and run businesses with the sole goal of increasing your net worth."
@@ -132,7 +130,7 @@ def load_variables(config_file="config.yaml"):
             ai_goals.append(ai_goal)
         if len(ai_goals) == 0:
             ai_goals = ["Increase net worth", "Grow Twitter Account", "Develop and manage multiple businesses autonomously"]
-         
+
     # Save variables to yaml file
     config = {"ai_name": ai_name, "ai_role": ai_role, "ai_goals": ai_goals}
     with open(config_file, "w") as file:
@@ -158,22 +156,22 @@ def construct_prompt():
             Fore.GREEN,
             f"Would you like me to return to being {config.ai_name}?",
             speak_text=True)
-        should_continue = input(f"""Continue with the last settings? 
+        should_continue = input(f"""Continue with the last settings?
 Name:  {config.ai_name}
 Role:  {config.ai_role}
-Goals: {config.ai_goals}  
+Goals: {config.ai_goals}
 Continue (y/n): """)
         if should_continue.lower() == "n":
             config = AIConfig()
 
-    if not config.ai_name:         
+    if not config.ai_name:
         config = prompt_user()
         config.save()
 
     # Get rid of this global:
     global ai_name
     ai_name = config.ai_name
-    
+
     full_prompt = config.construct_full_prompt()
     return full_prompt
 
@@ -234,7 +232,7 @@ def parse_arguments():
     global cfg
     cfg.set_continuous_mode(False)
     cfg.set_speak_mode(False)
-    
+
     parser = argparse.ArgumentParser(description='Process arguments.')
     parser.add_argument('--continuous', action='store_true', help='Enable Continuous Mode')
     parser.add_argument('--speak', action='store_true', help='Enable Speak Mode')
@@ -260,90 +258,90 @@ def parse_arguments():
 
 
 # TODO: fill in llm values here
+if __name__ == "__main__":
+    cfg = Config()
+    parse_arguments()
+    ai_name = ""
+    prompt = construct_prompt()
+    # print(prompt)
+    # Initialize variables
+    full_message_history = []
+    result = None
+    # Make a constant:
+    user_input = "Determine which next command to use, and respond using the format specified above:"
 
-cfg = Config()
-parse_arguments()
-ai_name = ""
-prompt = construct_prompt()
-# print(prompt)
-# Initialize variables
-full_message_history = []
-result = None
-# Make a constant:
-user_input = "Determine which next command to use, and respond using the format specified above:"
+    # Interaction Loop
+    while True:
+        # Send message to AI, get response
+        with Spinner("Thinking... "):
+            assistant_reply = chat.chat_with_ai(
+                prompt,
+                user_input,
+                full_message_history,
+                mem.permanent_memory,
+                cfg.fast_token_limit) # TODO: This hardcodes the model to use GPT3.5. Make this an argument
 
-# Interaction Loop
-while True:
-    # Send message to AI, get response
-    with Spinner("Thinking... "):
-        assistant_reply = chat.chat_with_ai(
-            prompt,
-            user_input,
-            full_message_history,
-            mem.permanent_memory,
-            cfg.fast_token_limit) # TODO: This hardcodes the model to use GPT3.5. Make this an argument
+        # print("assistant reply: "+assistant_reply)
+        # Print Assistant thoughts
+        print_assistant_thoughts(assistant_reply)
 
-    # print("assistant reply: "+assistant_reply)
-    # Print Assistant thoughts
-    print_assistant_thoughts(assistant_reply)
+        # Get command name and arguments
+        try:
+            command_name, arguments = cmd.get_command(assistant_reply)
+        except Exception as e:
+            print_to_console("Error: \n", Fore.RED, str(e))
 
-    # Get command name and arguments
-    try:
-        command_name, arguments = cmd.get_command(assistant_reply)
-    except Exception as e:
-        print_to_console("Error: \n", Fore.RED, str(e))
+        if not cfg.continuous_mode:
+            ### GET USER AUTHORIZATION TO EXECUTE COMMAND ###
+            # Get key press: Prompt the user to press enter to continue or escape
+            # to exit
+            user_input = ""
+            print_to_console(
+                "NEXT ACTION: ",
+                Fore.CYAN,
+                f"COMMAND = {Fore.CYAN}{command_name}{Style.RESET_ALL}  ARGUMENTS = {Fore.CYAN}{arguments}{Style.RESET_ALL}")
+            print(
+                "Enter 'y' to authorise command or 'n' to exit program...",
+                flush=True)
+            while True:
+                console_input = input(Fore.MAGENTA + "Input:" + Style.RESET_ALL)
+                if console_input.lower() == "y":
+                    user_input = "GENERATE NEXT COMMAND JSON"
+                    break
+                elif console_input.lower() == "n":
+                    user_input = "EXIT"
+                    break
+                else:
+                    continue
 
-    if not cfg.continuous_mode:
-        ### GET USER AUTHORIZATION TO EXECUTE COMMAND ###
-        # Get key press: Prompt the user to press enter to continue or escape
-        # to exit
-        user_input = ""
-        print_to_console(
-            "NEXT ACTION: ",
-            Fore.CYAN,
-            f"COMMAND = {Fore.CYAN}{command_name}{Style.RESET_ALL}  ARGUMENTS = {Fore.CYAN}{arguments}{Style.RESET_ALL}")
-        print(
-            "Enter 'y' to authorise command or 'n' to exit program...",
-            flush=True)
-        while True:
-            console_input = input(Fore.MAGENTA + "Input:" + Style.RESET_ALL)
-            if console_input.lower() == "y":
-                user_input = "GENERATE NEXT COMMAND JSON"
+            if user_input != "GENERATE NEXT COMMAND JSON":
+                print("Exiting...", flush=True)
                 break
-            elif console_input.lower() == "n":
-                user_input = "EXIT"
-                break
-            else:
-                continue
 
-        if user_input != "GENERATE NEXT COMMAND JSON":
-            print("Exiting...", flush=True)
-            break
+            print_to_console(
+                "-=-=-=-=-=-=-= COMMAND AUTHORISED BY USER -=-=-=-=-=-=-=",
+                Fore.MAGENTA,
+                "")
+        else:
+            # Print command
+            print_to_console(
+                "NEXT ACTION: ",
+                Fore.CYAN,
+                f"COMMAND = {Fore.CYAN}{command_name}{Style.RESET_ALL}  ARGUMENTS = {Fore.CYAN}{arguments}{Style.RESET_ALL}")
 
-        print_to_console(
-            "-=-=-=-=-=-=-= COMMAND AUTHORISED BY USER -=-=-=-=-=-=-=",
-            Fore.MAGENTA,
-            "")
-    else:
-        # Print command
-        print_to_console(
-            "NEXT ACTION: ",
-            Fore.CYAN,
-            f"COMMAND = {Fore.CYAN}{command_name}{Style.RESET_ALL}  ARGUMENTS = {Fore.CYAN}{arguments}{Style.RESET_ALL}")
+        # Execute command
+        if command_name.lower() != "error":
+            result = f"Command {command_name} returned: {cmd.execute_command(command_name, arguments)}"
+        else:
+            result = f"Command {command_name} threw the following error: " + arguments
 
-    # Execute command
-    if command_name.lower() != "error":
-        result = f"Command {command_name} returned: {cmd.execute_command(command_name, arguments)}"
-    else:
-        result = f"Command {command_name} threw the following error: " + arguments
-
-    # Check if there's a result from the command append it to the message
-    # history
-    if result is not None:
-        full_message_history.append(chat.create_chat_message("system", result))
-        print_to_console("SYSTEM: ", Fore.YELLOW, result)
-    else:
-        full_message_history.append(
-            chat.create_chat_message(
-                "system", "Unable to execute command"))
-        print_to_console("SYSTEM: ", Fore.YELLOW, "Unable to execute command")
+        # Check if there's a result from the command append it to the message
+        # history
+        if result is not None:
+            full_message_history.append(chat.create_chat_message("system", result))
+            print_to_console("SYSTEM: ", Fore.YELLOW, result)
+        else:
+            full_message_history.append(
+                chat.create_chat_message(
+                    "system", "Unable to execute command"))
+            print_to_console("SYSTEM: ", Fore.YELLOW, "Unable to execute command")
