+++ autogpt/token_counter.py
@@ -1,5 +1,7 @@
+from typing import Dict, List
+
 import tiktoken
-from typing import List, Dict
+
 
 def count_message_tokens(messages : List[Dict[str, str]], model : str = "gpt-3.5-turbo-0301") -> int:
     """
